adding APIManager and connect it in code
