+++ .gitignore
@@ -92,4 +92,7 @@ Pods
 # After new code Injection tools there's a generated folder /iOSInjectionProject
 # https://github.com/johnno1962/injectionforxcode
 
+# Keychain
+MarvelHeroesApp/API/Keychain
+
 iOSInjectionProject/
